Creates an MD5 hash. Use of Optional<String>
Demonstrates creating an MD5 hash of a string variable. Creates a string version of the hash using the statement "BigInteger(1, md5.digest()).toString(16)".  Uses the method 'ifPresent(Consumer<? super T> action)' of the Java Class 'Optional' to determine whether the result is null, passing the result to a 'Consumer<String>' for printing if a value is present.