Init version. Trevithick Day
Development of code for Trevithick Day

The first Trevithick Day was held in 1984 and since then it has been an annual fixture attracting thousands of visitors. It celebrates the huge contribution made by Richard Trevithick to Cornwall's industrial heritage.

Every year on the last Saturday in April, the streets of Camborne are filled with bal maidens and steam engines for the annual Trevithick Day celebrations. Trevithick Day is a free one-day festival celebrating the town's prominent industrial history and the life of local engineer Richard Trevithick.